+++ lizard_wms/models.py
@@ -1,5 +1,7 @@
 """Models for lizard_wms"""
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
+from urllib import urlencode
+import cgi
 import json
 import logging
 
@@ -12,7 +14,7 @@ from lizard_maptree.models import Category
 import owslib.wms
 import requests
 
-
+FIXED_WMS_API_VERSION = '1.1.1'
 logger = logging.getLogger(__name__)
 
 
@@ -22,8 +24,12 @@ class WMSConnection(models.Model):
     title = models.CharField(max_length=100)
     slug = models.CharField(max_length=100)
     url = models.URLField(verify_exists=False)
-    version = models.CharField(max_length=20, default='1.3.0',
-                               help_text=u"Version number for Lizard.")
+    version = models.CharField(
+        max_length=20,
+        default='1.3.0',
+        help_text=(
+            u"Version number for WMS service. Not used. 1.1.1 is used " +
+            u"because owslib can only handle 1.1.1."))
 
     params = models.TextField(
         default='{"height": "256", "width": "256", "layers": "%s", '
@@ -55,9 +61,12 @@ overwrites.""")
 
         if self.xml:
             xml = self.xml.encode('utf8').strip()
-            wms = owslib.wms.WebMapService(self.url, xml=xml)
+            wms = owslib.wms.WebMapService(self.url,
+                                           xml=xml,
+                                           version=FIXED_WMS_API_VERSION)
         else:
-            wms = owslib.wms.WebMapService(self.url)
+            wms = owslib.wms.WebMapService(self.url,
+                                           version=FIXED_WMS_API_VERSION)
 
         fetched = set()
         for name, layer in wms.contents.iteritems():
@@ -66,7 +75,8 @@ overwrites.""")
                 if layer.layers:
                     # Meta layer, don't use
                     continue
-                name = name.split(':', 1)[-1]  # owslib prepends with 'workspace:'.
+                name = name.split(':', 1)[-1]
+                # ^^^ owslib prepends with 'workspace:'.
                 kwargs = {'connection': self,
                           'name': name}
                 try:
@@ -78,7 +88,8 @@ overwrites.""")
                 layer_style = layer.styles.values()
                 # Not all layers have a description/legend.
                 if len(layer_style):
-                    layer_instance.description = '<img src="%s" alt="%s" />' % (
+                    tag = '<img src="%s" alt="%s" />'
+                    layer_instance.description = tag % (
                         layer_style[0]['legend'],
                         layer_style[0]['title'])
                 else:
@@ -111,6 +122,30 @@ overwrites.""")
                 layer.delete()
                 num_deleted += 1
         return num_deleted
+
+    def capabilities_url(self):
+        """Return the capabilities URL.
+
+        Copy/pasted mostly from owslib/wms.py. Only used in the admin for
+        debugging purposes.
+
+        """
+        qs = []
+        if self.url.find('?') != -1:
+            qs = cgi.parse_qsl(self.url.split('?')[1])
+
+        params = [x[0] for x in qs]
+
+        if 'service' not in params:
+            qs.append(('service', 'WMS'))
+        if 'request' not in params:
+            qs.append(('request', 'GetCapabilities'))
+        if 'version' not in params:
+            qs.append(('version', FIXED_WMS_API_VERSION))
+
+        urlqs = urlencode(tuple(qs))
+        return self.url.split('?')[0] + '?' + urlqs
+
 
 
 class WMSSource(models.Model):